requirements.txt
python-dotenv==1.0.0
Django==4.1.8
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
django-cors-headers==3.14.0
django-phone-field==1.8.1
Pillow==9.5.0
cloudinary==1.32.0
django-cloudinary-storage==0.3.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
whitenoise==6.4.0
